Packaging Component Updates - Added radio buttons for boxType selection, listed item limits for packaging per boxType.
The packit4me API used to obtain packaging configuration for boxes has a box + item limit of 500. For full functionality of this api, we also need to input every single possible box size for each item selected. These item limits are the result of meeting the requirements of our third party API.