Shipments Component work - sorting / saving options upon delete
The shipments table was not properly sorting by date, this was fixed by basing the date values upon unix time, which is attained by running the moment package to parse the current dates by running the moment package to parse the dates. The delete function now takes inthe current options on the table as an arguments and uses them as props for the rendering of the next table, which is then put in state upon the table's componentDidMount.