Shipment table modals - UI/UX experience for adding tracking numbers polished
Table now reliably refreshes upon adding tracking information, and forwards the user to the TRACKED SHIPMENTS list. Ideally, the last added shipment should appear at the top of the list - but a few shipments seemed to be added with an incorrect time 12 hours in the future. This may be due to herokuapps time data, and not an error on the front-end. Besides these incorrectly timed shipments, all newly added shipments are appearing at the top of the list as expected